Rename pages dir to views
