+++ src/views/city-map.vue
