+++ src/views/splash.vue
